Add resource to acquire and release locks
So that we can coordinate service restarts
across clusters to ensure that only one
node can restart at a time.

TODO - there are still a few more things to understand:

1. the failure cases. especially cases where we want to
acquire the lock, but we already have it, and the case
where we try to release a lock that we don't have.

2. I am not sure when we should release locks (ie:
   if lock release should be idempotent). Failures of
   any of these resources probably needs to be cleaned
   up manually until we better understand these operations.